==> disaster-tweets/01-keyword-analysis.ipynb <==
import pandas as pd

df_train = pd.read_csv('./train-dataset.csv')

keywords = df_train['keyword'].unique()
keywords[0:10]

keywords_count = df_train.groupby(['keyword']).size().reset_index(name='count')
keywords_count

keywords_count_per_target = df_train.groupby(['keyword', 'target']).size().reset_index(name='count').query('target == 1')
keywords_count_per_target.rename(columns={'count': 'positive_count'}, inplace=True)
keywords_count_per_target

keywords_stats = pd.merge(keywords_count, keywords_count_per_target, on='keyword', how='inner')
keywords_stats

keywords_stats['positive_factor'] = keywords_stats['positive_count'] / keywords_stats['count']
keywords_stats.drop(['count', 'target', 'positive_count'], axis=1, inplace=True)
keywords_stats

keywords_stats.to_csv('./keywords_stats.csv', index=False)
