==> disaster-tweets/02-urls_reference_analysis.ipynb <==
import requests
import validators as vld
import pandas as pd
import re
import urllib
import logging

df_train = pd.read_csv('./train.csv', index_col='id')
df_test = pd.read_csv('./test.csv', index_col='id')

urls = set()

def get_all_urls(source):
    for _, text in source['text'].items():
        tokens = re.split('\\s+', text)
        url_tokens = list(filter(vld.url, tokens))
        for url in url_tokens:
            urls.add(url)

get_all_urls(df_train)
get_all_urls(df_test)
print(len(urls))

import urllib3
urllib3.disable_warnings()

follow_domains = ['//t.co/', '//bit.ly/', '//dlvr.it/', 
                  '//tinyurl.com/', '//goo.gl/', '//ln.is/', '//shrd.by/', 
                  '//ift.tt/' , '//fb.me/',
                  '//on.doi.gov/', '//ow.ly/']

def follow_link(url, level=1):
    if level > 3:
        return [url]
    try:
        response = requests.get(url.replace('http://', 'https://'), verify=False, allow_redirects=False)
    except:
        return [url]
    if 'Location' in response.headers:
        url = response.headers['Location']
        for domain in follow_domains:
            if domain in url:
                out = follow_link(url, level+1)
                if not isinstance(out, (list, tuple)):
                    out = [out]
                elif isinstance(out, tuple):
                    out = list(out)
                if not url in out:
                    out.append(url)
                return out
        return [url]
    else:
        return [url]



result_url_expanded = []
for url in urls:
    output = follow_link(url)
    domain = ''
    last_url = None
    count = 1
    if output and isinstance(output, list):
        count = len(output)
        last_url = output[0]
    else:
        count = 1
        last_url = url
    domain = urllib.parse.urlparse(last_url).netloc
    result_url_expanded.append((url,domain,output,count))


df_url_domains = pd.DataFrame(data=result_url_expanded, columns=['url', 'final_url_domain', 'all_urls', 'url_redirects'])


df_url_domains.sample(n=20)

final_domain = []
for _, row in df_url_domains.iterrows():
    full_domain = row['final_url_domain']
    parts = full_domain.split('.')
    if len(parts) >= 2 and parts[-1] == 'uk' and parts[-2] == 'co':
        parts = parts[:-2] + ['co.uk']
    domain = '.'.join(parts[-2:])
    final_domain.append(domain)

final_domain[0:200:10]

df_url_domains['final_domain'] = final_domain

df_url_domains.to_csv('./url_stats.csv', index=False, columns=['url', 'final_url_domain', 'final_domain', 'url_redirects'])
