==> disaster-tweets/03-hashtags-analysis.ipynb <==
# Hashtags analysis

import pandas as pd
import numpy as np
import re
import string

pd.set_option('display.max_colwidth', 150)

df_train = pd.read_csv('./train-dataset.csv')
negative_count = sum(df_train['target'] == 0)
positive_count = sum(df_train['target'] == 1)
n_samples = df_train.shape[0]
negative_prop = negative_count / n_samples
positive_prop = positive_count / n_samples 
print(f'All samples: {n_samples}, negative: {100*negative_prop:.1f}%, positive: {100*positive_prop:.1f}%')


### Hypotesis - specific tags are more frequent for disaster tweets

hashtag_pattern = re.compile('#\w+')

def extract_hashtags(text):
    tokens = re.findall(hashtag_pattern, text)
    found_hashtags = filter(lambda token: len(token) > 0 and token[0]=='#', tokens)
    found_hashtags = list(map(lambda tag: tag.lstrip('#').lower(), found_hashtags))
    return found_hashtags


hashtags_series = []
all_hashtags = set()
PUNCTUATION = string.punctuation
hashtags_positive = {}
hashtags_negative = {}
for _, row in df_train.iterrows():
    text = row['text']
    target = int(row['target'])
    found_hashtags = extract_hashtags(text)
    for h in found_hashtags:
        all_hashtags.add(h)
    hashtags_series.append(found_hashtags)
    output = hashtags_positive if target else hashtags_negative
    for hashtag in found_hashtags:
        if hashtag in output:
            output[hashtag] = output[hashtag] + 1
        else:
            output[hashtag] = 1

df_train['hashtags'] = hashtags_series


df_train[['text', 'hashtags', 'target']].sample(n=10)

Lets calculate following metrics for hashcodes:
- `cnt_positive` - how many tweets labeled as disaster use the hash code.
- `cnt_negative` - how many tweets labeled as non-disaster use the hash code.
- `all_count` - all tweets having the hashcode.
- `positive_fact` - probability of using a hashtag used in disaster tweet.
- `negative_fact` - probability of using a hashtag in non-disaster tweet.
- `sentiment` - score [-1..1], scaled from `positive_fact` and an indication in which type of tweet the hashtag is used. Zero refers to no usage, or no bias.

# fill with zeros for not found hashtags
hashtag_counts = []
for h in all_hashtags:
    cnt_positive = hashtags_positive[h] if h in hashtags_positive else 0
    cnt_negative = hashtags_negative[h] if h in hashtags_negative else 0
    cnt_positive_adj = cnt_positive / positive_prop
    cnt_negative_adj = cnt_negative / negative_prop
    all_count = cnt_positive + cnt_negative
    positive_fact = cnt_positive / all_count if all_count > 0 else None
    sentiment = 2*(positive_fact - 0.5)
    sentiment_adj = (2*(cnt_positive_adj/(cnt_negative_adj+cnt_positive_adj)) - 1.0) if all_count > 0 else .0
    hashtag_counts.append((h, cnt_positive, cnt_negative, all_count, positive_fact, sentiment, sentiment_adj))

df_hashtags = pd.DataFrame(data=hashtag_counts, columns=['hashtag', 'cnt_positive', 'cnt_negative', 'all_count', 'positive_fact', 'sentiment', 'sentiment_adj']).set_index('hashtag')
df_hashtags.sort_values('all_count', ascending=False, inplace=True)

df_hashtags.count()

df_hashtags.query('all_count <3').count()

#### Most specific hash tags (with absolut sentiment value >= 0.5)

search_query = 'all_count >= 5 and (abs(sentiment) >= 0.5)'
df_hashtags.query(search_query)

df_hashtags.query('all_count > 10').sort_values('sentiment_adj', ascending=False).head(n=20)

df_hashtags.query('all_count > 10').sort_values('sentiment_adj', ascending=True).head(n=20)

#### How many tweets with most specific hash tags are used in?

query_cnt = df_hashtags.query(search_query)['all_count'].sum()
all_tweets = df_hashtags['all_count'].sum()
print(f'All tweets: {all_tweets}, tweets with relevant hashtags: {query_cnt} ({100*query_cnt/all_tweets:.0f}%)')

## Correlation with the target

hashtags_sentiment_result = []
hashtags_sentiment_adj_result = []
for _, row in df_train.iterrows():
    hashtags = row['hashtags']
    hashtag_sentiments = []
    hashtag_sentiments_adj = []
    weights = []
    for h in hashtags:
        counts = df_hashtags.loc[h]['all_count']
        hashtag_sentiments.append(df_hashtags.loc[h]['sentiment'])
        hashtag_sentiments_adj.append(df_hashtags.loc[h]['sentiment_adj'])
        weights.append(counts)
    sentiment = np.average(hashtag_sentiments, weights=weights) if len(hashtag_sentiments) > 0 else .0
    sentiment_adj = np.sum(hashtag_sentiments_adj) if len(hashtag_sentiments_adj) > 0 else .0
    hashtags_sentiment_result.append(sentiment)
    hashtags_sentiment_adj_result.append(sentiment_adj)
df_train['hashtags_sentiment'] = hashtags_sentiment_result
df_train['hashtags_sentiment_adj'] = hashtags_sentiment_adj_result

df_train.sample(n=10)

df_train[['target', 'hashtags_sentiment']].corr()


**Conflusion: there is many hashtags used more often for disaster tweets than for normal tweets.**


df_hashtags.sample(n=10)

df_hashtags.to_csv('./hashtags_sentiment.csv')
