==> disaster-tweets/20-enrich-datasets.ipynb <==
# Enrich test and training datasets

import numpy as np
import pandas as pd
import scipy as sp
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn import feature_extraction, ensemble, model_selection, pipeline, compose, preprocessing
from keyword_transformer import KeywordTransformer
from location_transformer import LocationTransformer
from text_stats_transformer import TextStatsTransformer


df_train = pd.read_csv('./train.csv', index_col='id')
df_test = pd.read_csv('./test.csv', index_col='id')


transformer = compose.ColumnTransformer(transformers=[
    ('keywords', KeywordTransformer(keyword_file_name='./keywords_stats.csv'), ['keyword']),
    ('location', LocationTransformer(), ['location']),
    ('text-features', TextStatsTransformer(), ['text'])
], remainder='passthrough', verbose_feature_names_out=False, verbose=True)

train_enriched = transformer.fit_transform(df_train)

df_train_enriched = pd.DataFrame(data = train_enriched, columns=transformer.get_feature_names_out())

transformer.get_feature_names_out()

df_train_enriched.to_csv('./train_enriched.csv', index_label='id')

test_enriched = transformer.fit_transform(df_test)
df_test_enriched = pd.DataFrame(data = test_enriched, columns=transformer.get_feature_names_out())
df_test_enriched.to_csv('./test_enriched.csv', index_label='id')
