==> disaster-tweets/21-text-embedding.ipynb <==
# Generate tweets text embeddings

import numpy as np
import pandas as pd
from sentence_transformers import SentenceTransformer


df_train = pd.read_csv('./train_enriched.csv', index_col='id')
df_train['text_content'].head()

model = SentenceTransformer('avsolatorio/GIST-small-Embedding-v0')

# Sentences we want to encode. Example:
sentence = ['This framework generates embeddings for each input sentence']

# Sentences are encoded by calling model.encode()
embedding = model.encode(sentence)

embedding

np.linalg.norm(embedding[0])

len(embedding[0])

def to_batches(lst, batch_size):
    for i in range(0, len(lst), batch_size):
        yield lst[i:i + batch_size]

def infer_embedding(data):
    batches = to_batches(data, batch_size=64)
    text_embedded = []
    for batch in batches:    
        text_embedded.extend(model.encode(batch))
    return text_embedded

text_embedded = infer_embedding(df_train['text_content'].to_list(), batch_size=64)
len(text_embedded), len(text_embedded[0])

import pickle

with open('./train-text-embeddings.pkl', 'wb') as fout:
    pickle.dump(text_embedded, fout)

df_test = pd.read_csv('./test_enriched.csv', index_col='id')
df_test['text_content'].head()

test_text_embedded = infer_embedding(df_test['text_content'].to_list())
len(test_text_embedded), len(test_text_embedded[0])

with open('./test-text-embeddings.pkl', 'wb') as fout:
    pickle.dump(test_text_embedded, fout)
